add update endpoint, add  extra configurstions

+++ controllers/post.controller.js
@@ -16,7 +16,7 @@ const getPosts = async (req, res) => {
 };
 
 const filterPosts = async (req, res) => {
-  if (!req.body.email) return res.status(400).send({ message: 'email user no provided' });
+  if (!req.body.email) return res.status(400).send({ message: "email user doesn't provided" });
 
   const user = await userService.getUser({ email: req.body.email });
 
@@ -35,9 +35,9 @@ const createPost = async (req, res) => {
   const { content, title } = req.body;
   const file = req.file;
 
-  if (!content) return res.status(400).send({ message: 'content no provided' });
-  if (!title) return res.status(400).send({ message: 'title no provided' });
-  if (!file) return res.status(400).send({ message: 'post image no provided' });
+  if (!content) return res.status(400).send({ message: "content doesn't provided" });
+  if (!title) return res.status(400).send({ message: "title doesn't provided" });
+  if (!file) return res.status(400).send({ message: "post image doesn't provided" });
 
   const imageUrl = await awsService.uploadFileAws(file);
 
@@ -54,12 +54,31 @@ const createPost = async (req, res) => {
   });
 };
 
+const updatePost = async (req, res) => {
+  const updates = {};
+
+  const postId = req.body.postId;
+
+  if (!postId) return res.status(400).send({ message: "postId doesn't provided" });
+
+  if (req.body.title) updates.title = req.body.title;
+  if (req.body.content) updates.content = req.body.content;
+
+  let result = '';
+  try {
+    result = await Post.findOneAndUpdate({ _id: postId, user: req.userId }, updates);
+  } catch (error) {
+    res.status(400).send({ message: 'Nothing updated' });
+  }
+  res.status(201).send({ message: result });
+};
+
 const deletePost = async (req, res) => {
   const { postId } = req.body;
   if (!postId) return res.status(400).send({ message: 'postId property no provided' });
 
   try {
-    await Post.findByIdAndDelete({ _id: postId });
+    await Post.findByIdAndDelete({ _id: postId, user: req.userId });
   } catch (error) {
     res.status(400).send({ message: "Post doesn't exist" });
   }
@@ -87,7 +106,7 @@ const getFilters = async (req, user) => {
   if (title) filters.title = title;
   if (content) filters.content = content;
 
-  filters.user = user.id;
+  if (user) filters.user = user.id;
 
   return filters;
 };
@@ -96,5 +115,6 @@ module.exports = {
   getPosts,
   filterPosts,
   createPost,
-  deletePost
+  deletePost,
+  updatePost
 };
